Implement VFolderPermissionTag.tsx
